Change/swap Join channel and join chat
This is a proposal that change the behavior of the Channel Row.
The Join arrow is removed and a chat bubble is added instead.

So now you press the text of the channel to join it (cvh.itemView.setOnClickListener), and the Chat bubble to join the Chat channel.

If someone want a more safe join channel, an option in seeting called something like "safe join channel" could be added, and a long press would be needed (cvh.itenView.senOnLongClickListener joins channel instead of opening the moreMenu)