Fix for JoinChannel scrolls channel list to button Issue#178
(e.g. use public server Libre Antenne)
Using the release candidate from Fdroid 3.3.0rc1:
When joining a channel on a larger server the display scrolls to another place on the list.
So it can be hard to find the group you joined.
The version 3.2.0 in playstore (with the old JoinChannel function) does not have this issue.

Changing mChannelView.smoothScrollToPosition to mChannelView.scrollToPosition (in ChannelListFragment.java->scrollToChannel) solves the problem.

Are there any reasons for using the smoothScrollToPosition instead og scrollToPosition? and do you think it´s because of a certain SDK version?
Otherwise I propose to change it.